[D3D9] Implemented combined texture-sampler and static sampler support in D3D9 backend.

+++ sources/Renderer/Direct3D9/RenderState/D3D9ProgrammablePSO.cpp
@@ -55,7 +55,7 @@ void D3D9ProgrammablePSO::Bind(D3D9StateManager& stateMngr)
     device->SetVertexShader(GetVertexShader());
     device->SetPixelShader(GetPixelShader());
 
-    stateMngr.SetStreamSourceFreqInstanceData(streamSourceFreq_.size(), streamSourceFreq_.data());
+    stateMngr.SetStreamSourceFreqInstanceData(static_cast<UINT>(streamSourceFreq_.size()), streamSourceFreq_.data());
 }
